completed

+++ controllers/admin.js
@@ -1,185 +1,186 @@
-const Product = require('../models/product');
+const Product = require("../models/product");
+const fileHeler = require("../util/file");
 
-const { validationResult } = require('express-validator');
-const mongoose = require('mongoose');
-
+const { validationResult } = require("express-validator");
+const mongoose = require("mongoose");
 
 exports.getAddProduct = (req, res, next) => {
-  res.render('admin/edit-product', {
-    pageTitle: 'Add Product',
-    path: '/admin/add-product',
+  res.render("admin/edit-product", {
+    pageTitle: "Add Product",
+    path: "/admin/add-product",
     editing: false,
     isAuthenticated: req.session.user,
     validationErrors: [],
-    errorMessage: req.flash('error'),
-    hasError: false
+    errorMessage: req.flash("error"),
+    hasError: false,
   });
 };
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
 
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-
-    return res.render('admin/edit-product', {
-      pageTitle: 'Add Product',
-      path: '/admin/add-product',
+    return res.render("admin/edit-product", {
+      pageTitle: "Add Product",
+      path: "/admin/add-product",
       editing: false,
       isAuthenticated: req.session.user,
       errorMessage: errors.array()[0].msg,
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
+        image: image,
         price: price,
-        description: description
+        description: description,
       },
-      validationErrors: errors.array()
+      validationErrors: errors.array(),
     });
-
   }
   const product = new Product({
     // _id: new mongoose.Types.ObjectId('644d74881307dc7b6c23769c'),
     title: title,
     price: price,
     description: description,
-    imageUrl: imageUrl,
-    userId: req.session.user
+    imageUrl: "/" + image.path,
+    userId: req.session.user,
   });
 
-  product.save()
-  .then(result => {
-    console.log(result);
-    res.redirect('/admin/products');
-  }).catch(err => {
-    // console.log(err)
-    // console.log("An Error Occured!")
+  product
+    .save()
+    .then((result) => {
+      console.log(result);
+      res.redirect("/admin/products");
+    })
+    .catch((err) => {
+      // console.log(err)
+      // console.log("An Error Occured!")
 
-    // res.status(500).render('admin/edit-product', {
-    //   pageTitle: 'Add Product',
-    //   path: '/admin/add-product',
-    //   editing: false,
-    //   isAuthenticated: req.session.user,
-    //   errorMessage: "An Error Occured",
-    //   hasError: true,
-    //   product: {
-    //     title: title,
-    //     imageUrl: imageUrl,
-    //     price: price,
-    //     description: description
-    //   },
-    //   validationErrors: []
-    // });
+      // res.status(500).render('admin/edit-product', {
+      //   pageTitle: 'Add Product',
+      //   path: '/admin/add-product',
+      //   editing: false,
+      //   isAuthenticated: req.session.user,
+      //   errorMessage: "An Error Occured",
+      //   hasError: true,
+      //   product: {
+      //     title: title,
+      //     imageUrl: imageUrl,
+      //     price: price,
+      //     description: description
+      //   },
+      //   validationErrors: []
+      // });
 
-    // res.redirect('/500');
+      // res.redirect('/500');
 
-    const error = new Error(err);
-    error.httpStatusCode = 500;
-    return next(error);
-  })
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.getEditProduct = (req, res, next) => {
   const editMode = req.query.edit;
   if (!editMode) {
-    return res.redirect('/');
+    return res.redirect("/");
   }
   const prodId = req.params.productId;
   Product.findById(prodId)
-    .then(product => {
-    if (!product) {
-      return res.redirect('/');
-    }
-    res.render('admin/edit-product', {
-      pageTitle: 'Edit Product',
-      path: '/admin/edit-product',
-      editing: editMode,
-      product: product,
-      isAuthenticated: req.session.user,
-      validationErrors: [],
-      hasError: false,
-      errorMessage: req.flash('error')
+    .then((product) => {
+      if (!product) {
+        return res.redirect("/");
+      }
+      res.render("admin/edit-product", {
+        pageTitle: "Edit Product",
+        path: "/admin/edit-product",
+        editing: editMode,
+        product: product,
+        isAuthenticated: req.session.user,
+        validationErrors: [],
+        hasError: false,
+        errorMessage: req.flash("error"),
+      });
+    })
+    .catch((err) => {
+      console.log(err);
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
     });
-  }).catch(err => {
-    console.log(err)
-    const error = new Error(err);
-    error.httpStatusCode = 500;
-    return next(error);
-  });
 };
 
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
 
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    return res.render('admin/edit-product', {
-      pageTitle: 'Edit Product',
-      path: '/admin/edit-product',
+    return res.render("admin/edit-product", {
+      pageTitle: "Edit Product",
+      path: "/admin/edit-product",
       editing: false,
       isAuthenticated: req.session.user,
       errorMessage: errors.array()[0].msg,
       hasError: true,
       product: {
         prodId: prodId,
-        title:  updatedTitle,
+        title: updatedTitle,
         price: updatedPrice,
-        imageUrl: updatedImageUrl,
-        description:  updatedDesc,
-        _id: prodId
+        description: updatedDesc,
+        _id: prodId,
       },
-      validationErrors: errors.array()
+      validationErrors: errors.array(),
     });
   }
 
-  Product.findById(prodId).then(product => {
-
-    if(product.userId.toString() !== req.user._id.toString()){
-      return res.redirect('/');
-    }
-    product.title = updatedTitle;
-    product.price = updatedPrice;
-    product.imageUrl = updatedImageUrl;
-    product.description = updatedDesc;
-   return product.save()
-  })
-  .then(
-    result => {
-      console.log('PRODUCT UPDATED SUCESSFULLY!!');
-      res.redirect('/admin/products');
-    }
-  )
-  .catch(err => {
-    console.log(err)
-    const error = new Error(err);
-    error.httpStatusCode = 500;
-    return next(error);
-    
-  })
+  Product.findById(prodId)
+    .then((product) => {
+      if (product.userId.toString() !== req.user._id.toString()) {
+        return res.redirect("/");
+      }
+      product.title = updatedTitle;
+      product.price = updatedPrice;
+      if (image) {
+        fileHeler.deleteFile(product.imageUrl);
+      }
+      product.imageUrl = image ? "/" + image.path : product.imageUrl;
+      product.description = updatedDesc;
+      return product.save();
+    })
+    .then((result) => {
+      console.log("PRODUCT UPDATED SUCESSFULLY!!");
+      res.redirect("/admin/products");
+    })
+    .catch((err) => {
+      console.log(err);
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find({userId: req.user._id})
+  Product.find({ userId: req.user._id })
     // .select('title price -_id') // Use to filter result
     // .populate('userId') //eagerload the relationship
     // .populate('userId', 'name -_id') //eagerload the relationship with filter
-    .then(products => {
-      console.log('products',products)
-      res.render('admin/products', {
+    .then((products) => {
+      console.log("products", products);
+      res.render("admin/products", {
         prods: products,
-        pageTitle: 'Admin Products',
-        path: '/admin/products',
-        isAuthenticated: req.session.user
+        pageTitle: "Admin Products",
+        path: "/admin/products",
+        isAuthenticated: req.session.user,
       });
-    }).catch(err => {
+    })
+    .catch((err) => {
       const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
@@ -189,12 +190,20 @@ exports.getProducts = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
   // Product.findByIdAndRemove(prodId)
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-  .then(
-    result => {
-      console.log('PRODUCT DELETED SUCESSFULLY!!');
-      res.redirect('/admin/products');
-    }
-  )
-  .catch(err => console.log(err));
+
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found!"));
+      }
+
+      fileHeler.deleteFile(product.imageUrl);
+
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
+    .then((result) => {
+      console.log("PRODUCT DELETED SUCESSFULLY!!");
+      res.redirect("/admin/products");
+    })
+    .catch((err) => console.log(err));
 };
